+++ src/knex.js
@@ -15,7 +15,7 @@ export function getKnex() {
     connection: process.env.PG_CONNECTION_STRING,
     pool: {
       min: 0,
-      max: 50,
+      max: 20,
     },
   });
